Stop unnecessarily using uint64 variable for uint32 data
UserHistoryPredictor::LearningSegmentFingerprint returns uint32 data.  We don't need to use set<uint64> there.

No behavior change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@334 a6090854-d499-a067-5803-1114d4e51264

+++ src/prediction/user_history_predictor.cc
@@ -965,7 +965,7 @@ bool UserHistoryPredictor::LookupEntry(
     string key = entry->key();
     string value = entry->value();
     const Entry *current_entry = entry;
-    set<uint64> seen;
+    set<uint32> seen;
     seen.insert(EntryFingerprint(*current_entry));
     // Until target entry gets longer than input_key.
     while (key.size() <= input_key.size()) {
